+++ ChangeLog.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - support for multipart requests to :ring module and improved the request/response
   dump functions
 - module :http-client-legacy. A simple HTTP client based on the JDK HttpUrlConnection
+- function `bytebuf-merge` to merge byte buffers
 
 ### Fixed 
 - the return status handling in the :ring module